requirements.txt
tensorflow
numpy
opencv-python
matplotlib
flask
opendatasets
scipy
pillow
scikit-learn 
facenet-pytorch
sort-tracker-py
